+++ sdks/python/apache_beam/typehints/typed_pipeline_test.py
@@ -63,7 +63,7 @@ class MainInputTest(unittest.TestCase):
       [1, 2, 3] | beam.Map(str.upper)
 
   def test_loose_bounds(self):
-    @typehints.with_input_types(typehints.Union[int, float, int])
+    @typehints.with_input_types(typehints.Union[int, float, long])
     @typehints.with_output_types(basestring)
     def format_number(x):
       return '%g' % x
